# api-docs.yaml
openapi: 3.0.0
info:
  title: Leads Core API
  description: |
    Комплексное API для управления виджетами лидогенерации.
    
    Сервис предоставляет возможности для:
    - Создания и управления виджетами
    - Сбора форм и событий от виджетов
    - Получения статистики и аналитики
    - Мониторинга состояния системы
    
    ## Аутентификация
    
    API использует JWT токены для аутентификации. Токен должен быть передан в заголовке Authorization:
    ```
    Authorization: Bearer <jwt_token>
    ```
    
    ## Пагинация
    
    Многие эндпоинты поддерживают пагинацию с параметрами:
    - `page` - номер страницы (по умолчанию 1)
    - `per_page` или `limit` - количество элементов на странице (по умолчанию 20, максимум 100)
    
    ## Коды ответов
    
    - `200` - Успешный запрос
    - `201` - Ресурс создан
    - `204` - Успешный запрос без содержимого
    - `400` - Ошибка валидации или неверный запрос
    - `401` - Не авторизован
    - `403` - Доступ запрещен
    - `404` - Ресурс не найден
    - `405` - Метод не поддерживается
    - `500` - Внутренняя ошибка сервера
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/ad/leads-core
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.leads.example.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/v1/widgets:
    get:
      tags:
        - Widgets
      summary: Получить список виджетов пользователя
      description: Возвращает список виджетов текущего пользователя с поддержкой пагинации
      parameters:
        - name: page
          in: query
          description: Номер страницы
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          description: Количество элементов на странице
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Список виджетов
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      widgets:
                        type: array
                        items:
                          $ref: '#/components/schemas/Widget'
                      meta:
                        $ref: '#/components/schemas/Meta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags:
        - Widgets
      summary: Создать новый виджет
      description: Создает новый виджет для текущего пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWidgetRequest'
      responses:
        '201':
          description: Виджет создан
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Widget'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/widgets/{id}:
    get:
      tags:
        - Widgets
      summary: Получить виджет по ID
      description: Возвращает информацию о конкретном виджете
      parameters:
        - name: id
          in: path
          required: true
          description: Уникальный идентификатор виджета
          schema:
            type: string
      responses:
        '200':
          description: Информация о виджете
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Widget'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Widgets
      summary: Обновить виджет
      description: Обновляет настройки существующего виджета
      parameters:
        - name: id
          in: path
          required: true
          description: Уникальный идентификатор виджета
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWidgetRequest'
      responses:
        '200':
          description: Виджет обновлен
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Widget'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags:
        - Widgets
      summary: Удалить виджет
      description: Удаляет виджет и все связанные с ним данные
      parameters:
        - name: id
          in: path
          required: true
          description: Уникальный идентификатор виджета
          schema:
            type: string
      responses:
        '204':
          description: Виджет удален
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен для аутентификации
  schemas:
...
    Widget:
      type: object
      properties:
        id:
          type: string
          description: Уникальный идентификатор виджета
          example: "4be0643f-1d98-573b-97cd-ca98a65347dd"
        owner_id:
          type: string
          description: ID владельца виджета
          example: "user_456def"
        type:
          type: string
          description: Тип виджета
          example: "lead-form"
          enum: ["lead-form", "banner", "action", "social-proof", "live-interest", "widget-tab", "sticky-bar", "quiz", "wheelOfFortune"]
        name:
          type: string
          description: Название виджета
          example: "Форма обратной связи"
        isVisible:
          type: boolean
          description: Активен ли виджет
          example: true
        config:
           $ref: '#/components/schemas/WidgetConfig'
        created_at:
          type: string
          format: date-time
          description: Время создания
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Время последнего обновления
          example: "2024-01-16T14:20:00Z"
        stats:
          $ref: '#/components/schemas/WidgetStats'
    WidgetConfig:
      type: object
...
    WidgetStats:
      type: object
      properties:
        views:
          type: integer
          format: int64
          description: Количество просмотров
          example: 1254
        submits:
          type: integer
          format: int64
          description: Количество отправок
          example: 87
        closes:
          type: integer
          format: int64
          description: Количество закрытий
          example: 156
        last_view:
          type: string
          format: date-time
          description: Время последнего просмотра
          example: "2024-01-16T15:30:00Z"
...
    CreateWidgetRequest:
      type: object
      required:
        - type
        - name
        - isVisible
        - config
      properties:
        type:
          type: string
          description: Тип виджета
          example: "lead-form"
          enum: ["lead-form", "banner", "action", "social-proof", "live-interest", "widget-tab", "sticky-bar", "quiz", "wheelOfFortune"]
        name:
          type: string
          description: Название виджета
          example: "Форма обратной связи"
          minLength: 1
          maxLength: 255
        isVisible:
          type: boolean
          description: Активен ли виджет
          example: true
        config:
          type: object
          description: Настройки полей виджета
    UpdateWidgetRequest:
      type: object
      properties:
        name:
          type: string
          description: Название виджета
          example: "Новое название"
          minLength: 1
          maxLength: 100
        isVisible:
          type: boolean
          description: Активен ли виджет
          example: false
...
    Meta:
      type: object
      properties:
        page:
          type: integer
          description: Текущая страница
          example: 1
        per_page:
          type: integer
          description: Элементов на странице
          example: 20
        total:
          type: integer
          description: Общее количество элементов
          example: 156
        has_more:
          type: boolean
          description: Есть ли еще страницы
          example: true
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Описание ошибки
          example: "Validation failed"
        details:
          description: Дополнительные детали ошибки
          example:
            field: "email"
            message: "Invalid email format"
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: "Validation error"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "email"
              message:
                type: string
                example: "Invalid email format"

    # Health & Metrics
...
  responses:
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "User not found"
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Widget not found"
    ValidationError:
      description: Ошибка валидации
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            error: "Validation error"
            details:
              - field: "email"
                message: "Invalid email format"
              - field: "name"
                message: "Name is required"
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Internal server error"
...